pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mvp-auth"
version = "0.1.0"
description = "Pluggable async JWT authentication library for rapid MVP development"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0,<3.0",
    "pydantic-settings>=2.0,<3.0",
    "pyjwt>=2.0,<3.0",
    "bcrypt>=4.0,<5.0",
]

[project.optional-dependencies]
argon2 = ["argon2-cffi>=23.1"]
postgres = [
    "sqlalchemy[asyncio]>=2.0,<3.0",
    "asyncpg>=0.29",
]
fastapi = ["fastapi>=0.100"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
]
all = ["mvp-auth[argon2,postgres,fastapi,dev]"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
